+++ R/http--request.r
@@ -6,12 +6,6 @@ make_request <- function(method, handle, url, config = NULL, body = NULL,
   stopifnot(is.handle(handle))
   stopifnot(is.character(url), length(url) == 1)
 
-  # Combine with default config
-  opts <- modify_config(default_config(), config)
-  if (method != "POST") {
-    opts$customrequest <- method
-  }
-
   # Sign request, if needed
   token <- opts$token
   if (!is.null(token)) {