+++ R/request.R
@@ -1,15 +1,20 @@
-request <- function(method = "GET", url = NA_character_, headers = character(),
-                    fields = list(),
-                    options = character(), auth_token = NULL,
-                    output = c("memory", "file", "stream")) {
-  stopifnot(is.character(method), length(method) == 1)
-  stopifnot(is.character(url), length(url) == 1)
-  stopifnot(is.character(headers))
-  output <- match.arg(output)
+request <- function(method = NULL, url = NULL, headers = NULL,
+                    fields = NULL, options = NULL, auth_token = NULL,
+                    output = NULL) {
+  if (!is.null(method))
+    stopifnot(is.character(method), length(method) == 1)
+  if (!is.null(url))
+    stopifnot(is.character(url), length(url) == 1)
+  if (!is.null(headers))
+    stopifnot(is.character(headers))
+  if (!is.null(fields))
+    stopifnot(is.list(fields))
+  if (!is.null(output))
+    stopifnot(inherits(output, "write_function"))
 
   structure(
     list(
-      method = toupper(method),
+      method = method,
       url = url,
       headers = keep_last(headers),
       fields = fields,
@@ -20,68 +25,69 @@ request <- function(method = "GET", url = NA_character_, headers = character(),
     class = "request"
   )
 }
+is.request <- function(x) inherits(x, "request")
 
 request_default <- function() {
-  cert <- system.file("cacert.pem", package = "httr")
   request(
     options = list(
-      followlocation = TRUE,
-      maxredirs = 10L,
-      accept_encoding = "gzip",
       useragent = default_ua(),
-      if (.Platform$OS.type == "windows") config(cainfo = cert),
+      cainfo = find_cert_bundle(),
       getOption("httr_config")
     ),
-    headers = c(Accept = "application/json, text/xml, application/xml, */*")
+    headers = c(Accept = "application/json, text/xml, application/xml, */*"),
+    output = write_function("write_memory")
   )
 }
 
-is.request <- function(x) inherits(x, "request")
+as.request <- function(x) UseMethod("as.request")
+as.request.list <- function(x) structure(x, class = "request")
+as.request.request <- function(x) x
 
-combine_requests <- function(x, y) {
+request_build <- function(method, url, config = list(), ...) {
+  extra <- list(...)
+  extra[has_names(extra)] <- NULL
+
+  req <- Reduce(request_combine, extra)
+  if (!identical(config, list))
+    req <- request_combine(as.request(config), req)
+
+  req$method <- method
+  req$url <- url
+
+  req
+}
+
+request_combine <- function(x, y) {
+  if (length(x) == 0 && length(y) == 0) return(request())
+  if (length(x) == 0) return(y)
+  if (length(y) == 0) return(x)
   stopifnot(is.request(x), is.request(y))
 
   request(
-    y$method,
-    y$url,
-    keep_last(x$headers, y$headers),
-    c(x$fields, y$fields),
-    keep_last(x$options, y$options),
-    y$output
+    method =     y$method %||% x$method,
+    url =        y$url %||% x$url,
+    headers =    keep_last(x$headers, y$headers),
+    fields =     c(x$fields, y$fields),
+    options =    keep_last(x$options, y$options),
+    auth_token = y$auth_token %||% x$auth_token,
+    output =     y$output %||% x$output
   )
-}
-
-#' @export
-c.request <- function(...) {
-  Reduce(combine_requests, list(...))
 }
 
 #' @export
 print.request <- function(x, ...) {
-  x <- request_build(x)
-
-  cat(x$method, " ", x$url, " -> ", x$output, "\n", sep = "")
+  cat("<request>\n")
+  if (!is.null(x$method) && !is.null(x$url))
+    cat(toupper(x$method), " ", x$url, "\n", sep = "")
+  if (!is.null(x$output))
+  cat("Output: ", class(x$output)[[1]], "\n", sep = "")
   named_vector("Options", x$options)
   named_vector("Headers", x$headers)
   named_vector("Fields", x$fields)
 }
 
-named_vector <- function(title, x) {
-  if (length(x) == 0) return()
-
-  cat(title, ":\n", sep = "")
-  bullets <- paste0("* ", names(x), ": ", as.character(x))
-  cat(bullets, sep = "\n")
-}
-
-keep_last <- function(...) {
-  x <- c(...)
-  x[!duplicated(names(x), fromLast = TRUE)]
-}
-
-
-request_build <- function(req) {
-  req <- c(request_default(), req)
+request_prepare <- function(req) {
+  req <- request_combine(request_default(), req)
 
   if (req$method != "POST")
     req$options$customrequest <- req$method
@@ -93,22 +99,20 @@ request_build <- function(req) {
   req
 }
 
-# Abstract over the differences in RCurl API depending on whether or not
-# you send a body.
-httr_perform <- function(req, handle) {
+request_perform <- function(req, handle) {
   stopifnot(is.request(req), inherits(handle, "curl_handle"))
-  req <- request_build(req)
+  req <- request_prepare(req)
 
-  # Set handle options
-  do.call(curl::handle_setopt, c(list(handle), req$options))
-  do.call(curl::handle_setheaders, c(list(handle), req$headers))
-  do.call(curl::handle_setfields, c(list(handle), req$fields))
+  browser()
+  curl::handle_setopt(handle, .list = req$options)
+  curl::handle_setheaders(handle, .list = req$headers)
+  curl::handle_setform(handle, .list = req$fields)
   on.exit(curl::handle_reset(handle), add = TRUE)
 
-  resp <- curl::curl_perform(req$url, handle)
+  resp <- request_fetch(req$output, req$url, handle)
 
-  all_headers <- curl::parse_headers(resp$headers, multiple = TRUE)
-  headers <- last(all_headers)
+  all_headers <- parse_headers(resp$headers)
+  headers <- last(all_headers)$headers
   if (!is.null(headers$date)) {
     date <- parse_http_date(headers$Date)
   } else {
@@ -116,14 +120,14 @@ httr_perform <- function(req, handle) {
   }
 
   response(
-    url = res$url,
-    status_code = res$status_code,
+    url = resp$url,
+    status_code = resp$status_code,
     headers = headers,
     all_headers = all_headers,
     cookies = curl::handle_cookies(handle),
-    content = content,
+    content = resp$content,
     date = date,
-    times = res$times,
+    times = resp$times,
     request = request
   )
 }