+++ R/url.r
@@ -118,7 +118,7 @@ build_url <- function(url) {
   if (is.list(url$query)) {
     url$query <- compact(url$query)
     names <- curl::curl_escape(names(url$query))
-    values <- curl::curl_escape(url$query)
+    values <- curl::curl_escape(as.character(url$query))
 
     query <- paste0(names, "=", values, collapse = "&")
   } else {